+++ src/pyunigps/unitypes_core.py
@@ -3,8 +3,8 @@ UNI Protocol core globals, constants, datatypes and message identifiers.
 
 Created on 26 Jan 2026
 
-Information sourced from public domain Unicore UM980 Interface Specifications © 2023, Unicore
-https://www.ardusimple.com/wp-content/uploads/2023/04/Unicore-Reference-Commands-Manual-For-N4-High-Precision-Products_V2_EN_R1.4-1.pdf
+Information sourced from public domain Unicore Reference Commands R1.13 © Dec 2025 Unicore
+https://en.unicore.com/uploads/file/Unicore%20Reference%20Commands%20Manual%20For%20N4%20High%20Precision%20Products_V2_EN_R1.13.pdf
 
 :author: semuadmin (Steve Smith)
 """