+++ src/pyunigps/unitypes_poll.py
@@ -3,9 +3,6 @@ UNI Protocol query definitions.
 
 Created on 26 Jan 2026
 
-Information sourced from public domain Unicore UM980 Interface Specifications © 2023, Unicore
-https://www.ardusimple.com/wp-content/uploads/2023/04/Unicore-Reference-Commands-Manual-For-N4-High-Precision-Products_V2_EN_R1.4-1.pdf
-
 :author: semuadmin (Steve Smith)
 """